Support transfer sizes of over wMaxPacketSize
The API exposed by embassy-usb assumes that each call sends or
receives one packet, not one transfer. In order to handle transfers
larger than wMaxPacketSize, we must add logic to the app looping until
we receive a transfer that is smaller than wMaxPacketSize, which
terminates a USB transfer.

While at that, also increase USB buffer to 4K as we have a lot of free
memory still.

Signed-off-by: Felipe Balbi <febalbi@microsoft.com>